Split out the stdlib monkeypatches to a different file
